+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added new warning labels
 * Added new reviews
 * Added Inhaler
+* Added notification on first beating the game
 
 # v1.0